requirements.txt
# Core Framework
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic>=2.5
python-multipart>=0.0.9

# Data Processing
pandas>=2.0
numpy>=1.24
openpyxl>=3.1.0
xlrd>=2.0.1  # For .xls files
pyxlsb>=1.0.10  # For .xlsb files
pyarrow>=14.0.0  # For .parquet files
fastparquet>=2023.10.0  # Alternative parquet support

# Machine Learning
scikit-learn>=1.4
joblib>=1.3

# Logging
loguru>=0.7.0

# Document Processing (for policy management)
python-docx>=1.1.0
PyPDF2>=3.0.0

# Utilities
python-dateutil>=2.8.0

# Database Support - دعم قواعد البيانات
pyodbc>=5.0.0  # SQL Server ODBC driver
sqlalchemy>=2.0.0  # SQL toolkit and ORM
pymssql>=2.2.0  # Alternative SQL Server driver (pure Python)